Use Paper's plugin API
I'm aware that this might break for some people that aren't using
Paper, which is why I won't be creating a pull request against the
main repository.

We need to use Paper's plugin API as it introduces support for the
PlayerPickupExperienceEvent which we need to keep track of the origin
of Experience Orbs.